Split the TestStand modules into specific modules for pin mapping and driver specification
