+++ examples/nidcpower_source_dc_voltage/_nidcpower_helpers.py
@@ -4,19 +4,19 @@ from typing import Any, Dict, Optional
 
 import grpc
 import nidcpower
-from _constants import USE_SIMULATION
 
 import ni_measurementlink_service as nims
 
 
 def create_session(
     session_info: nims.session_management.SessionInformation,
+    use_simulation: bool,
     session_grpc_channel: Optional[grpc.Channel] = None,
     initialization_behavior: nidcpower.SessionInitializationBehavior = nidcpower.SessionInitializationBehavior.AUTO,
 ) -> nidcpower.Session:
     """Create driver session based on reserved session and grpc channel."""
     options: Dict[str, Any] = {}
-    if USE_SIMULATION:
+    if use_simulation:
         options["simulate"] = True
         options["driver_setup"] = {"Model": "4141"}
