+++ examples/nidcpower_source_dc_voltage/measurement.py
@@ -78,7 +78,9 @@ def measure(
         grpc_device_channel = get_grpc_device_channel(
             measurement_service, nidcpower, service_options
         )
-        with create_session(reservation.session_info, grpc_device_channel) as session:
+        with create_session(
+            reservation.session_info, service_options.use_simulation, grpc_device_channel
+        ) as session:
             channels = session.channels[reservation.session_info.channel_list]
             channel_mappings = reservation.session_info.channel_mappings
